+++ src/pyupway/config/__init__.py
@@ -1,0 +1,1 @@
+from .myupwayconfig import MyUpwayConfig