+++ src/pyupway/exceptions/responseerror.py
@@ -1,0 +1,4 @@
+class ResponseError(Exception):
+    def __init__(self, message="Error with the server response"):
+        self.message = message
+        super().__init__(self.message)