+++ src/pyupway/models/variablevalue.py
@@ -1,0 +1,10 @@
+from dataclasses import dataclass
+from ..enums import Variable
+
+@dataclass
+class VariableValue:
+    Id: int
+    Name: str
+    Enumerator: Variable
+    Value: int | float | str | bool | None
+    Unit: str | None