+++ src/pyupway/services/__init__.py
@@ -1,0 +1,2 @@
+from .myuplinkservice import MyUplinkService
+from .myupwayservice import MyUpwayService