+++ changelog.md
@@ -1,3 +1,11 @@
+# v0.0.8
+
+- ADD: Integration for MyUplink service
+
+## Other Stuff
+
+- This release includes breaking changes! Basically all the previous MyUpway functionality is there, but you need to update the configurations. Make sure you define the DataService you want to use and be specific of what you specify.
+
 # v0.0.7
 
 - ADD: Smart Price enums